restbucks/app/serializers.py
from rest_framework import serializers

from app.models import Product


class ProductSerializer(serializers.ModelSerializer):
	class Meta:
		model = Product
		exclude = ['created_at', 'last_modified', 'slug']
