restbucks/requirements.txt
Django==1.11.24
pytz==2019.2
sqlparse==0.3.0
djangorestframework
